code refractored
