+++ content-scripts/utils.js
