+++ popup/index.html
@@ -44,6 +44,6 @@
             <span class="light_text">Version <span id="ext_version"></span></span>
         </div>
     </div>
-    <script src="popup.js"></script>
+    <script src="popup.js" type="module"></script>
 </body>
 </html>
